% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{add_counting_process}
\alias{add_counting_process}
\title{Add Counting Process (Step Plot)}
\usage{
add_counting_process(H_t, T_max, col = "black")
}
\arguments{
\item{H_t}{Numeric vector of event timestamps.}

\item{T_max}{The end of the observation window.}

\item{col}{Colour for the step line. Default is base R's default black.}
}
\description{
Layers a step-function representing the counting process N(t).
}
