% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{add_events}
\alias{add_events}
\title{Add Event Ticks to Plot}
\usage{
add_events(H_t, col = "red")
}
\arguments{
\item{H_t}{Numeric vector of event timestamps.}

\item{col}{Tick colour. Default is base R's default red.}
}
\description{
Adds rug ticks representing event times.
}
