% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{add_intensity}
\alias{add_intensity}
\title{Add Intensity Line to Plot}
\usage{
add_intensity(t, intensities, col = "black")
}
\arguments{
\item{t}{Numeric vector of time points (the grid).}

\item{intensities}{Numeric vector of intensity values.}

\item{col}{Line colour. Default is base R's default black.}
}
\description{
Layers an intensity curve onto an existing plot.
}
