% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helpers.R
\name{diagnose_sim_study}
\alias{diagnose_sim_study}
\title{Plot MCMC diagnostics for a simulation study replicate}
\usage{
diagnose_sim_study(study, R, start = 1, thin = 1, params = "all", ...)
}
\arguments{
\item{study}{List; output from \code{run_simulation_study}.}

\item{R}{Integer; number of individual replicate to diagnose.}

\item{start}{Integer; first iteration of interest.}

\item{thin}{Integer; the required interval between successive samples.}

\item{...}{optional arguments to pass to MCMCvis::MCMCtrace}
}
\description{
Plot MCMC diagnostics for a simulation study replicate
}
