% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-functions.R
\name{.new_point_process_sim}
\alias{.new_point_process_sim}
\title{Internal Constructor for point_process_sim}
\usage{
.new_point_process_sim(
  events,
  T_max,
  params,
  type,
  marks = NULL,
  full_history = NULL
)
}
\description{
Internal Constructor for point_process_sim
}
\keyword{internal}
