% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{exp_mhp_compensator}
\alias{exp_mhp_compensator}
\title{Exponential Marked Hawkes Process Compensator}
\usage{
exp_mhp_compensator(theta, H_t, marks, t)
}
\arguments{
\item{theta}{Numeric vector of parameters \eqn{c(\lambda_0, A, \beta)}.}

\item{H_t}{Numeric vector of event timestamps.}

\item{marks}{Numeric vector of marks corresponding to \code{H_t}.}

\item{t}{Numeric value. The end of the observation window.}
}
\description{
Calculates the integrated ground-process intensity over
\eqn{[0,t]}.
}
