% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{exp_mhp_fit}
\alias{exp_mhp_fit}
\title{Fit Marked Hawkes Process via MLE}
\usage{
exp_mhp_fit(
  object,
  marks = NULL,
  T_max = NULL,
  lik_method = c("fast", "slow"),
  init = c(0.5, 1, 1)
)
}
\arguments{
\item{object}{Either a numeric vector of event times or a
\code{marked_pp_sim} object.}

\item{marks}{Numeric vector of event marks when \code{object} is a numeric vector.
Not required when \code{object} is a \code{marked_pp_sim}.}

\item{T_max}{Numeric value. Required if \code{object} is a numeric vector.}

\item{lik_method}{Character string, either \code{"fast"} or \code{"slow"}.}

\item{init}{Numeric vector of initial values for
\eqn{c(\lambda_0,A,\beta)}.}
}
\value{
An object of class \code{marked_hawkes_fit}.
}
\description{
Estimates \eqn{\lambda_0}, \eqn{A}, and \eqn{\beta} by maximum
likelihood for a marked Hawkes process with exponential temporal decay
and linear mark productivity.
}
