% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{exp_mhp_loglik}
\alias{exp_mhp_loglik}
\title{Marked Hawkes Log-Likelihood (Exponential)}
\usage{
exp_mhp_loglik(theta, H_t, marks, T_max, lik_method = c("fast", "slow"))
}
\arguments{
\item{theta}{Numeric vector of parameters \eqn{c(\lambda_0, A, \beta)}.}

\item{H_t}{Numeric vector of event timestamps.}

\item{marks}{Numeric vector of marks corresponding to \code{H_t}.}

\item{T_max}{Numeric value. The end of the observation window.}

\item{lik_method}{Character string, either \code{"fast"} or \code{"slow"}.}
}
\description{
Calculates the ground-process log-likelihood for a marked
Hawkes process with excitation \eqn{A m \exp(-\beta t)}.
}
