% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marked-hawkes-npb.R
\name{logpost_kappa}
\alias{logpost_kappa}
\title{Log full conditional for the mark time recaling parameter.}
\usage{
logpost_kappa(
  kappa,
  marks,
  times,
  O_idx,
  z,
  s,
  theta,
  mu_kappa,
  sd_kappa,
  kernel
)
}
\arguments{
\item{kappa}{Numeric; proposed mark time recaling parameter.}

\item{marks}{Numeric vector; event marks.}

\item{times}{Numeric vector; event times.}

\item{O_idx}{Integer vector; offspring-event indices.}

\item{z}{Integer vector; parent indices.}

\item{s}{Integer vector; kernel-component allocations.}

\item{theta}{Numeric vector; kernel atom locations.}

\item{mu_kappa}{Numeric; Normal prior mean.}

\item{sd_kappa}{Numeric; Normal prior standard deviation.}

\item{kernel}{Character; \code{"step"} or \code{"pwlin"}.}
}
\value{
Numeric; log full conditional density at \code{kappa}.
}
\description{
Evaluates the log full conditional of \code{kappa}.
}
\keyword{internal}
