% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{plot_base_canvas}
\alias{plot_base_canvas}
\title{Create a Base R Canvas}
\usage{
plot_base_canvas(
  xlim,
  ylim,
  title,
  xlab = "Time (t)",
  ylab = expression(lambda^"*" * (t))
)
}
\arguments{
\item{xlim}{Numeric vector of length 2 (e.g., c(0, T)).}

\item{ylim}{Numeric vector of length 2 (e.g., c(0, max_y)).}

\item{title}{String for the plot title.}

\item{xlab}{String for the x-axis label. Default "Time (t)".}

\item{ylab}{String for the y-axis label. Default intensity symbol.}
}
\description{
Sets up a standard plot area using default base R styling.
}
