% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{residuals.marked_hawkes_fit}
\alias{residuals.marked_hawkes_fit}
\title{Calculate Transformed Residuals for a Marked Hawkes Fit}
\usage{
\method{residuals}{marked_hawkes_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{marked_hawkes_fit}.}

\item{...}{Additional arguments.}
}
\description{
Calculates time-rescaled residuals using the compensator of the
fitted ground process. For unpredictable marks, the mark density does not
enter the transformed event-time residuals.
}
