% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marked_hawkes_npb.R
\name{default_proposal_sds}
\alias{default_proposal_sds}
\title{Default Metropolis proposal standard deviations}
\usage{
default_proposal_sds()
}
\description{
Default Metropolis proposal standard deviations
}
\keyword{internal}
