% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marked-hawkes-npb.R
\name{update_zs}
\alias{update_zs}
\title{Sample latent parent and kernel-component allocations}
\usage{
update_zs(
  lambda0,
  A,
  beta = NULL,
  w,
  C,
  theta,
  times,
  marks,
  kernel = c("step", "pwlin"),
  mark_productivity = c("linear", "exponential"),
  kappa = 0,
  mark_time_rescaling = FALSE,
  theta_max = NULL
)
}
\arguments{
\item{lambda0}{Numeric; baseline intensity.}

\item{A}{Numeric; mark productivity scale.}

\item{beta}{Numeric; exponential mark-productivity slope, if required.}

\item{w}{Numeric vector; length-K mixture weights.}

\item{C}{Numeric; kernel normalising constant.}

\item{theta}{Numeric vector; length-K kernel atom locations.}

\item{times}{Numeric vector; sorted event times.}

\item{marks}{Numeric vector; event marks.}

\item{kernel}{Character; \code{"step"} or \code{"pwlin"}.}

\item{mark_productivity}{Character; \code{"linear"} or \code{"exponential"}.}

\item{kappa}{Numeric; value of mark time rescaling paramter. Default to zero
(no rescaling)}

\item{mark_time_rescaling}{whether to enable mark-dependent temporal
rescaling of the excitation kernel. Default \code{FALSE}..}

\item{theta_max}{Numeric; if not NULL then is the end of the finite support of the
Hawkes kernel, \eqn{[0, \theta_{max}]}}
}
\value{
List with integer vectors \code{z} (parent indices, with 0 for
immigrants) and \code{s} (kernel-component allocations).
}
\description{
Samples the parent event and mixture component for each event using
a categorical draw over all valid immigrant and offspring assignments.
}
\keyword{internal}
